pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "streamlit>=1.42.0",
]
